# cmd/api/feed.go
package main

import (
	"net/http"

	"github.com/DenysBahachuk/gopher_social/internal/store"
)

// getUserFeedHandler godoc
//
//	@Summary		Fetches the user feed
//	@Description	Fetches the user feed
//	@Tags			feed
//	@Accept			json
//	@Produce		json
//	@Param			since	query		string	false	"Since"
//	@Param			until	query		string	false	"Until"
//	@Param			limit	query		int		false	"Limit"
//	@Param			offset	query		int		false	"Offset"
//	@Param			sort	query		string	false	"Sort"
//	@Param			tags	query		string	false	"Tags"
//	@Param			search	query		string	false	"Search"
//	@Success		200		{object}	[]store.PostForFeed
//	@Failure		400		{object}	error
//	@Failure		500		{object}	error
//	@Security		ApiKeyAuth
//	@Router			/users/feed [get]
func (app *application) getUserFeedHandler(w http.ResponseWriter, r *http.Request) {
	pfq := store.PaginatedFeedQuery{
		Limit:  20,
		Offset: 0,
		Sort:   "desc",
	}

	err := pfq.Parse(r)
	if err != nil {
		app.badRequestErrorResponse(w, r, err)
		return
	}

	if err = Validate.Struct(pfq); err != nil {
		app.badRequestErrorResponse(w, r, err)
		return
	}

	feed, err := app.store.Posts.GetUserFeed(r.Context(), int64(130), pfq)
	if err != nil {
		app.internalServerErrorResponse(w, r, err)
		return
	}

	err = app.writeResponse(w, http.StatusOK, feed)
	if err != nil {
		app.internalServerErrorResponse(w, r, err)
		return
	}
}